+++ easydev/easydev/DataAttribute.h
@@ -5,13 +5,13 @@
 class DataAttribute
 {
 public:
-	DataAttribute(const std::string& name, const uint64_t& size);
+	DataAttribute(const std::string& name, const uint32_t& size);
 	virtual ~DataAttribute();
 
 	std::string GetName();
 private:
     std::string name_;
-	uint64_t size_;
+	uint32_t size_;
 	void* data_;
 };
